MyAutoTest/requirements.txt
selenium>=4.9.1
requests>=2.31.0
PyYAML>=5.4.1
PyMySQL>=1.1.0
pytest>=7.4.2
pytest-check>=2.1.4
allure-pytest>=2.13.2
pytest-result-log>=1.2.1
pytest-sugar>=0.9.7
playwright>=1.39.0
loguru>=0.4.1
Jinja2>=3.1.2
xlwings~=0.30.12
jsonpath~=0.82.2